+++ cleanup.sh
@@ -1,0 +1,1 @@
+find . -type d \( -name "__pycache__" -o -name ".pytest_cache" \) -exec rm -rf {} +
