+++ requirements.txt
@@ -8,7 +8,6 @@ httpcore==1.0.9
 httpx==0.28.1
 idna==3.10
 iniconfig==2.1.0
-logging==0.4.9.6
 packaging==25.0
 pluggy==1.5.0
 pydantic==2.11.4
@@ -20,6 +19,5 @@ python-dotenv==1.1.0
 redis==6.2.0
 sniffio==1.3.1
 starlette==0.46.2
-typing-inspection==0.4.0
 typing_extensions==4.13.2
 uvicorn==0.34.2
